psycho: start to decouple psycho_ctx usages
The current design is not modular by any stretch, and will
end up unmaintainable very quickly. For example, the CPU
interpreter needs to care about the whole context, rather
than just the CPU. This is not efficient.

This will take some time to finish, but is important to do
so before continuing ahead with further refactoring and
cleanups.

Signed-off-by: Michael Rodriguez <me@mcrod.dev>