Fix delay slot handling and branch address calcs
We now take a copy of the current PC, load the instruction there, and
then allow the "pipelining" to go on.

Fix branch target address calculations: the branch target is relative to
the address of the branch instruction itself, plus 4 to skip the branch
instruction.

These changes fix visual bugs in the disassembler, and logic errors in
the interpreter.